duet_monitor/requirements.txt
pyserial==3.5
pandas==2.2.1
matplotlib==3.8.3 